model/requirements.txt
torch
torchvision
transformers
opencv-python
pillow
fastapi
uvicorn
python-multipart